+++ js/renders.js
@@ -1,0 +1,5 @@
+export function resultToList (result) {
+  return result
+    .map(elem => `<li data-id="${elem.id}">${elem.title}</li>`)
+    .join('')
+}
