+++ js/sort.js
@@ -1,0 +1,7 @@
+export function sortByVote (a, b) {
+  return b.popularity * b.vote_average - a.popularity * a.vote_average
+}
+
+export function sortByPolularity (a, b) {
+  return b.popularity - a.popularity
+}
