+++ package.json
@@ -13,6 +13,7 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "api": "^1.0.0",
     "lodash": "^4.17.11",
     "nes.css": "^2.1.1"
   },
@@ -23,6 +24,7 @@
     "eslint-plugin-node": "^9.1.0",
     "eslint-plugin-promise": "^4.1.1",
     "eslint-plugin-standard": "^4.0.0",
+    "jest": "^24.8.0",
     "parcel-bundler": "^1.12.3"
   }
 }
